+++ src/Stack/FileWatch.hs
@@ -11,7 +11,7 @@ import Control.Concurrent.Async (race_)
 import Control.Concurrent.STM
 import Control.Exception (Exception)
 import Control.Exception.Enclosed (tryAny)
-import Control.Monad (forever, unless)
+import Control.Monad (forever, unless, when)
 import qualified Data.ByteString.Lazy as L
 import qualified Data.Map.Strict as Map
 import Data.Monoid ((<>))
@@ -37,6 +37,7 @@ fileWatch :: IO (Path Abs Dir)
           -> ((Set (Path Abs File) -> IO ()) -> IO ())
           -> IO ()
 fileWatch getProjectRoot inner = withManager $ \manager -> do
+    allFiles <- newTVarIO Set.empty
     dirtyVar <- newTVarIO True
     watchVar <- newTVarIO Map.empty
     projRoot <- getProjectRoot
@@ -48,10 +49,14 @@ fileWatch getProjectRoot inner = withManager $ \manager -> do
                  return $ FileIgnoredChecker (const False)
           Right chk -> return chk
 
-    let onChange = atomically $ writeTVar dirtyVar True
+    let onChange event = atomically $ do
+            files <- readTVar allFiles
+            when (eventPath event `Set.member` files) (writeTVar dirtyVar True)
 
         setWatched :: Set (Path Abs File) -> IO ()
         setWatched files = do
+            print files
+            atomically $ writeTVar allFiles $ Set.map toFilePath files
             watch0 <- readTVarIO watchVar
             let actions = Map.mergeWithKey
                     keepListening
@@ -77,7 +82,7 @@ fileWatch getProjectRoot inner = withManager $ \manager -> do
                 return Nothing
             startListening = Map.mapWithKey $ \dir () -> do
                 let dir' = fromString $ toFilePath dir
-                listen <- watchDir manager dir' (not . isFileIgnored . eventPath) (const onChange)
+                listen <- watchDir manager dir' (not . isFileIgnored . eventPath) onChange
                 return $ Just listen
 
     let watchInput = do
@@ -90,7 +95,7 @@ fileWatch getProjectRoot inner = withManager $ \manager -> do
                         putStrLn "quit: exit"
                         putStrLn "build: force a rebuild"
                         putStrLn "watched: display watched directories"
-                    "build" -> onChange
+                    "build" -> atomically $ writeTVar dirtyVar True
                     "watched" -> do
                         watch <- readTVarIO watchVar
                         mapM_ (putStrLn . toFilePath) (Map.keys watch)